Refactored package names, as well as updated rest endpoint
