+++ src/main/java/com/github/szilex94/edu/round_tracker/mappers/UserMapper.java
@@ -1,8 +1,8 @@
 package com.github.szilex94.edu.round_tracker.mappers;
 
-import com.github.szilex94.edu.round_tracker.repository.user.UserProfileDao;
-import com.github.szilex94.edu.round_tracker.rest.user.UserDto;
-import com.github.szilex94.edu.round_tracker.service.user.User;
+import com.github.szilex94.edu.round_tracker.repository.user.profile.UserProfileDao;
+import com.github.szilex94.edu.round_tracker.rest.user.profile.UserDto;
+import com.github.szilex94.edu.round_tracker.service.user.profile.User;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
