+++ src/main/java/com/github/szilex94/edu/round_tracker/rest/user/profile/UserDto.java
@@ -1,4 +1,4 @@
-package com.github.szilex94.edu.round_tracker.rest.user;
+package com.github.szilex94.edu.round_tracker.rest.user.profile;
 
 import lombok.Data;
 import lombok.experimental.Accessors;
